Refactor: Rename DialogComponent to SidePanel
Rename `RightSideDialog` to `SidePanel` and update related references.

* **Component Renaming**
  - Rename `RightSideDialog.razor` to `SidePanel.razor`
  - Update `@onclick` event handlers to use `sidePanelService.CloseSide()`
  - Update `id` attributes to `sidePanelCloseButton` and `sidePanelEditButton`
  - Update `@if` conditions to use `ShowClose` and `ShowEdit`

* **Class Renaming**
  - Rename `RightSideDialog` class to `SidePanel`
  - Update constructor to use `sidePanelService` instead of `dialogService`
  - Update methods to use `sidePanelService`

* **Service Updates**
  - Rename `RightSideDialog` references to `SidePanel` in `InnovativeDialogService.cs`
  - Update `OpenDynamicFormDialogWithOptions` method to use `SidePanel`
  - Update `GetFormTitle` and `GetWidth` methods to use `SidePanel` and `SidePanelWidth`

* **Example App Updates**
  - Rename `RightSideDialog` references to `SidePanel` in `InnovativeDialogServiceExample.razor` and `InnovativeDialogServiceExample.razor.cs`

* **Test Updates**
  - Rename `RightSideDialog` references to `SidePanel` in `DialogTests.cs`
  - Update test methods to use `SidePanel`

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/innovative365nl/sharedblazorcompoments?shareId=XXXX-XXXX-XXXX-XXXX).
